pyproject.toml
[project]
name = "agent-a-to-z"
version = "0.1.0"
description = "탄탄한 AI 에이전트 A to Z"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]==0.177.0",
    "dotenv>=0.9.9",
    "python-telegram-bot[job-queue]==22.3",
    "firecrawl-py==3.4.0",
    "feedparser==6.0.11",
    "python-dateutil>=2.9.0.post0",
]
